backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "google-auth>=2.41.1",
    "hypercorn>=0.17.3",
    "requests>=2.32.5",
]
